% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_coding.R
\name{add_color_code}
\alias{add_color_code}
\title{Title}
\usage{
add_color_code(table, col, bounds = NULL, colors = NULL)
}
\arguments{
\item{table}{A gt table}

\item{col}{The column to be color coded}

\item{bounds}{The bounds for the column being color coded}

\item{colors}{The color scale to apply}
}
\value{
a color coded gt table
}
\description{
Title
}
