% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_coding.R
\name{add_color_codes}
\alias{add_color_codes}
\title{Title}
\usage{
add_color_codes(table, cols, bounds_list = NULL, colors = NULL)
}
\arguments{
\item{table}{A gt table}

\item{cols}{The columns to be color coded}

\item{bounds_list}{A list of bounds for each column being color coded}

\item{colors}{The color scale to apply}
}
\value{
A color coded gt table
}
\description{
Title
}
