% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_format_table.R
\name{auto_format_table}
\alias{auto_format_table}
\title{Automatically formats the values in each column according to their data type}
\usage{
auto_format_table(gt_table)
}
\arguments{
\item{gt_table}{The static table}
}
\value{
A value formatted table
}
\description{
Automatically formats the values in each column according to their data type
}
\examples{
auto_format_table(summary_table)
}
