% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_themes.R
\name{available_themes}
\alias{available_themes}
\title{The list of available table themes}
\usage{
available_themes()
}
\value{
The themes list
}
\description{
The list of available table themes
}
\examples{
available_themes()
}
