% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theming.R
\name{dt_theme}
\alias{dt_theme}
\title{Provides custom theming functionality for DT tables}
\usage{
dt_theme(table, title, theme = "greyscale")
}
\arguments{
\item{table}{The DT table to be themed}

\item{theme}{The theme to apply to the table}
}
\value{
The themed DT table
}
\description{
Provides custom theming functionality for DT tables
}
\examples{
dt_theme(financial_data, title = '2023 Financial Earnings', theme = 'greenscale')
}
