% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouping.R
\name{group_table}
\alias{group_table}
\title{Change the table grouping}
\usage{
group_table(table, group_col = NULL)
}
\arguments{
\item{table}{Input table.}

\item{group_col}{The grouping column.}
}
\value{
The grouped table.
}
\description{
Change the table grouping
}
\examples{
group_table(gtcars_table, group_col = "mfr")
}
