% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecloth.R
\name{tablecloth}
\alias{tablecloth}
\title{This function takes a data frame as input and produces a stylized table
that can be static or dynamic}
\usage{
tablecloth(
  data,
  title = NULL,
  subtitle = NULL,
  rowname_col = NULL,
  group_col = NULL,
  aggregate_by = NULL,
  format = "static",
  theme = "default",
  ...
)
}
\arguments{
\item{data}{The data}

\item{title}{The title of your table}

\item{format}{The table format - 'static' or 'dynamic'}

\item{theme}{The table theme}

\item{...}{}
}
\value{
A formatted table
}
\description{
This function takes a data frame as input and produces a stylized table
that can be static or dynamic
}
\examples{
gs_tablecloth(sample_data, title = 'My Sample Data', format = 'static', theme = 'greyscale')
}
