% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecloth.R
\name{dynamic_table}
\alias{dynamic_table}
\title{Constructs the dynamic table with an optional theme using DT}
\usage{
dynamic_table(data, rowname_col, group_col, title, theme)
}
\arguments{
\item{data}{The table data}

\item{title}{The title of the table}

\item{theme}{The custom theme}
}
\value{
The formatted dynamic table
}
\description{
Constructs the dynamic table with an optional theme using DT
}
\examples{
dynamic_table(financial_data, title = '2023 Income Statement', theme = 'greenscale')
}
