% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecloth.R
\name{static_table}
\alias{static_table}
\title{Constructs the static table with an optional theme using gt}
\usage{
static_table(
  data,
  table_number = NULL,
  title = NULL,
  subtitle = NULL,
  rowname_col = NULL,
  group_col = NULL,
  aggregate_by = NULL,
  autoformat = TRUE,
  order_cols = NULL,
  rename_cols = NULL,
  hide_cols = NULL,
  is_color_coded = FALSE,
  theme = "default"
)
}
\arguments{
\item{data}{The table data}

\item{title}{The title for the table}

\item{subtitle}{The subtitle for the table}

\item{rowname_col}{The row name column}

\item{group_col}{The grouping column}

\item{autoformat}{Whether or not the values should be automatically formatted}

\item{theme}{The custom theme}
}
\value{
The formatted static table
}
\description{
Constructs the static table with an optional theme using gt
}
\examples{
static_table(financial_data, theme = 'greenscale')
}
