% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablecloth.R
\name{tablecloth}
\alias{tablecloth}
\title{This function takes a data frame as input and produces a stylized table
that can be static (gt) or dynamic (DT)}
\usage{
tablecloth(
  data,
  table_number = NULL,
  title = NULL,
  subtitle = NULL,
  rowname_col = NULL,
  group_col = NULL,
  aggregate_by = NULL,
  autoformat = TRUE,
  order_cols = NULL,
  rename_cols = NULL,
  hide_cols = NULL,
  is_color_coded = FALSE,
  theme = "default",
  format = "static"
)
}
\arguments{
\item{data}{The data}

\item{title}{The title of your table}

\item{theme}{The table theme}

\item{format}{The table format - 'gt' or 'dt'}
}
\value{
A formatted table
}
\description{
This function takes a data frame as input and produces a stylized table
that can be static (gt) or dynamic (DT)
}
\examples{
tablecloth(gtcars, format = 'dynamic', theme = 'greyscale')
}
